+++ core/src/main/kotlin/com/tdt4240/group3/model/MapGenerator.kt
@@ -1,15 +1,21 @@
-package com.tdt4240.group3.model.temporaryFactory
+package com.tdt4240.group3.model
 
 import com.badlogic.ashley.core.Engine
 import com.badlogic.ashley.core.Entity
-import com.tdt4240.group3.model.MapData
-import com.tdt4240.group3.model.ecs.components.CityComponent
-import com.tdt4240.group3.model.ecs.components.PositionComponent
-import com.tdt4240.group3.model.ecs.components.TeamComponent
-import com.tdt4240.group3.model.ecs.components.TileComponent
-import com.tdt4240.group3.model.team.TeamName
-import ktx.ashley.allOf
-import kotlin.math.abs
+import com.tdt4240.group3.config.MapData
+import com.tdt4240.group3.model.components.CityComponent
+import com.tdt4240.group3.model.components.PositionComponent
+import com.tdt4240.group3.model.components.TeamComponent
+import com.tdt4240.group3.model.components.TileComponent
+import com.tdt4240.group3.model.entities.CapitalConfig
+import com.tdt4240.group3.model.entities.CapitalFactory
+import com.tdt4240.group3.model.entities.CityConfig
+import com.tdt4240.group3.model.entities.CityFactory
+import com.tdt4240.group3.model.entities.TileConfig
+import com.tdt4240.group3.model.entities.TileFactory
+import com.tdt4240.group3.model.entities.TroopConfig
+import com.tdt4240.group3.model.entities.TroopFactory
+import com.tdt4240.group3.model.hexmap.MapCalculations
 import kotlin.math.floor
 import kotlin.random.Random
 
@@ -30,14 +36,14 @@ class MapGenerator(private val engine: Engine) {
     }
 
     //TODO("This function should probably be in a different file, maybe TurnSystem")
-    fun createTroopFromCity(cityEntity: Entity, troopType: String): Entity {
+    fun createTroopFromCity(cityEntity: Entity, unitType: UnitType): Entity {
         val city = CityComponent.Companion.mapper.get(cityEntity)
         val position = PositionComponent.Companion.mapper.get(cityEntity)
         val team = TeamComponent.Companion.mapper.get(cityEntity)
         return troopFactory.createEntity(
             TroopConfig(
                 team.team,
-                troopType,
+                unitType,
                 city.baseProduction,
                 position.q,
                 position.r
@@ -45,18 +51,13 @@ class MapGenerator(private val engine: Engine) {
         )
     }
 
-    //TODO("The variables should be changed for more clarity")
-    private fun hexDistance(q1: Int, r1: Int, q2: Int, r2: Int): Int {
-        return (abs(q1 - q2) + abs(q1 + r1 - q2 - r2) + abs(r1 - r2)) / 2
-    }
-
     fun generateCities(count: Int, capitalPositions: List<Pair<Int, Int>>) {
         val cityNames = MapData.CITY_NAMES.toMutableList()
-            .also { it.shuffle(Random(System.currentTimeMillis())) }
+            .also { it.shuffle(Random(42)) }
 
-        val tileFamily = allOf(PositionComponent::class, TileComponent::class).get()
+        val tileFamily = ktx.ashley.allOf(PositionComponent::class, TileComponent::class).get()
         val allTiles = engine.getEntitiesFor(tileFamily).map { entity ->
-            val pos = PositionComponent.Companion.mapper.get(entity)
+            val pos = PositionComponent.mapper.get(entity)
             Pair(pos.q, pos.r)
         }
 
@@ -64,42 +65,40 @@ class MapGenerator(private val engine: Engine) {
 
         val placedCities = mutableListOf<Pair<Int, Int>>()
 
-        val shuffled = candidateTiles.shuffled(Random(System.currentTimeMillis()))
+        val shuffled = candidateTiles.shuffled(Random(42))
         for (tile in shuffled) {
             if (placedCities.size >= count) break
             val tooClose = placedCities.any { placed ->
-                hexDistance(placed.first, placed.second, tile.first, tile.second) < 2
+                MapCalculations.hexDistance(placed.first, placed.second, tile.first, tile.second) < 2
             } || capitalPositions.any { cap ->
-                hexDistance(cap.first, cap.second, tile.first, tile.second) < 2
+                MapCalculations.hexDistance(cap.first, cap.second, tile.first, tile.second) < 2
             }
             if (!tooClose) {
                 placedCities.add(tile)
                 val name = cityNames.getOrElse(placedCities.size - 1) { "City ${placedCities.size}" }
-                cityFactory.createEntity(
-                    CityConfig(
-                        name = name,
-                        baseProduction = 10,
-                        q = tile.first,
-                        r = tile.second,
-                        team = TeamName.NONE
-                    )
-                )
+                cityFactory.createEntity(CityConfig(
+                    name = name,
+                    baseProduction = 10,
+                    q = tile.first,
+                    r = tile.second,
+                    team = Team.NONE
+                ))
             }
         }
     }
 
     //TODO("Should be refactored for better clarity")
-    fun generateCapitals(teams: List<TeamName>): List<Pair<Int, Int>> {
+    fun generateCapitals(teams: List<Team>): List<Pair<Int, Int>> {
         val capitalNames = MapData.CAPITAL_NAMES.toMutableList()
-            .also { it.shuffle(Random(System.currentTimeMillis())) }
+            .also { it.shuffle(Random(42)) }
 
-        val tileFamily = allOf(PositionComponent::class, TileComponent::class).get()
+        val tileFamily = ktx.ashley.allOf(PositionComponent::class, TileComponent::class).get()
 
         data class TilePos(val q: Int, val r: Int, val x: Float, val y: Float)
 
         val validTiles = engine.getEntitiesFor(tileFamily)
             .mapNotNull { entity ->
-                val pos = PositionComponent.Companion.mapper.get(entity) ?: return@mapNotNull null
+                val pos = PositionComponent.mapper.get(entity) ?: return@mapNotNull null
                 val coords = pos.q to pos.r
                 if (coords in MapData.WATER_TILES) return@mapNotNull null
                 TilePos(pos.q, pos.r, pos.x, pos.y)
@@ -152,19 +151,15 @@ class MapGenerator(private val engine: Engine) {
             used.add(coords)
             placedCapitals.add(coords)
 
-            capitalFactory.createEntity(
-                CapitalConfig(
-                    name = capitalNames.getOrElse(index) { "Capital ${index + 1}" },
-                    baseProduction = 20,
-                    q = bestTile.q,
-                    r = bestTile.r,
-                    team = team
-                )
-            )
+            capitalFactory.createEntity(CapitalConfig(
+                name = capitalNames.getOrElse(index) { "Capital ${index + 1}" },
+                baseProduction = 20,
+                q = bestTile.q,
+                r = bestTile.r,
+                team = team
+            ))
         }
 
         return placedCapitals
     }
-
-
 }
