+++ core/src/main/kotlin/com/tdt4240/group3/model/entities/Factory.kt
@@ -1,4 +1,4 @@
-package com.tdt4240.group3.model.temporaryFactory
+package com.tdt4240.group3.model.entities
 
 import com.badlogic.ashley.core.Entity
