+++ core/src/main/kotlin/com/tdt4240/group3/model/entities/GameStateFactory.kt
@@ -1,7 +1,7 @@
-package com.tdt4240.group3.model.temporaryFactory
+package com.tdt4240.group3.model.entities
 
 import com.badlogic.ashley.core.Engine
-import com.tdt4240.group3.model.ecs.components.GameStateComponent
+import com.tdt4240.group3.model.components.GameStateComponent
 import ktx.ashley.entity
 import ktx.ashley.with
