+++ core/src/main/kotlin/com/tdt4240/group3/model/hexmap/MapCalculations.kt
@@ -4,6 +4,7 @@ import kotlin.math.abs
 
 object MapCalculations {
 
+    //TODO("The variables should be changed for more clarity")
     fun hexDistance(q1: Int, r1: Int, q2: Int, r2: Int): Int {
         return (abs(q1 - q2) +
             abs(q1 + r1 - q2 - r2) +
